pyproject.toml
[build-system]
requires = ["setuptools", "versioneer[toml]", "tomli; python_version < '3.11'"]
build-backend = "setuptools.build_meta"

[project]
name = "picframe"
description = "Picture frame viewer powered by raspberry with homeassistant integration"
keywords = [
    "picframe", 
    "viewer", 
    "raspberry", 
    "raspi",
    "homeassistant", 
    "hass"]
authors = [
  { name = "Helge Erbe", email = "helge@erbehome.de"  },
  { name = "Paddy Gaunt" },
  { name = "Jeff Godfrey" },
]
maintainers = [
  { name = "Helge Erbe", email = "helge@erbehome.de"  },
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers=[
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Viewers",
]
dynamic = ["version"]
dependencies = [
    "Pillow>=10.2.0",
    "defusedxml",
    "pi3d>=2.54",
    "PyYAML",
    "paho-mqtt>=2.1.0",
    "IPTCInfo3",
    "numpy",
    "ninepatch>=0.2.0",
    "pi_heif>=0.8.0",
    "python-vlc",
    "rubicon-objc; sys_platform == 'darwin'"
]

[project.optional-dependencies]
test = [
    "pytest",
]

[project.urls]
"Homepage" = "https://github.com/helgeerbe/picframe"

[project.scripts]
picframe = "picframe.start:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]


[tool.setuptools.package-data]
"picframe.data" = ["**"]
"picframe.config" = ["**"]
"picframe.html" = ["**"]

[tool.setuptools.dynamic]
version = {attr = "picframe.__version__"}
readme = {file = ["README.rst", "USAGE.rst"]}

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "src/picframe/_version.py"
versionfile_build = "picframe/_version.py"
tag_prefix = ""
parentdir_prefix = "picframe-"
